Separate out categorical colors from brand colors
