+++ packages/component-library/src/_Themes/CategoricalColors.js
@@ -1,0 +1,28 @@
+export default {
+  pink: {
+    rgb: "rgb(220,69,86)",
+    rgba: "rgba(220,69,86,1)",
+    hex: "#DC4556"
+  },
+  // Called "teal" in style story
+  green: {
+    rgb: "rgb(25,183,170)",
+    rgba: "rgba(25,183,170,1)",
+    hex: "#19B7AA"
+  },
+  blue: {
+    rgb: "rgb(30,98,189)",
+    rgba: "rgba(30,98,189,1)",
+    hex: "#1E62BD"
+  },
+  purple: {
+    rgb: "rgb(114,29,124)",
+    rgba: "rgba(114,29,124,1)",
+    hex: "#721D7C"
+  },
+  yellow: {
+    rgb: "rgb(255,178,31)",
+    rgba: "rgba(255,178,38,1)",
+    hex: "#FFB226"
+  }
+};
